requirements.txt
google-generativeai>=0.5.0
requests>=2.31.0
rich>=13.0.0
PyPDF2>=3.0.0